Refactored the home page to use custom components.
